Minor rewrite of the code.
Revised SQLite queries and table schemas. Removed the archive_id field
from the jobs table. Added a query to delete a job that is used if it is
detected that a job no longer exists on Glacier.

Removed the Glacier class and moved its operation functions to the
top-level. Fixed bugs that some of them had, and added a new operation
to get job status(es). Removed the close() function since it seems
unnecessary.

Switched from using getop to argparse; still working the kinks out of
this model but overall it's more dependable and flexible. Rewrote the
logic for handling .boto config file loads and established a clearer
order of precedence.